+++ lib/src/extension/num.dart
@@ -4,19 +4,22 @@ extension DoubleExt on double {
   num toPrecision(int fractionDigits) {
     if (_canBeInt) return toInt();
     final mod = pow(10, fractionDigits.toDouble()).toDouble();
-    return (this * mod).round().toDouble() / mod;
+    final calculation = (this * mod).round().toDouble() / mod;
+    return calculation._canBeInt ? calculation.toInt() : calculation;
   }
 
   bool get _canBeInt => this % 1 == 0;
 
   num get toIntIfTrue => _canBeInt ? toInt() : this;
+
+  double get getRandomDouble => Random().nextDouble();
 }
 
 extension IntExt on int {
   String get bytesToBinaryPrefix => _formatBytes(this, binaryPrefixes: true);
   String get bytesToSIUnit => _formatBytes(this, binaryPrefixes: false);
 
-  int get getRandomNumberFromZeroToLessThanThis => Random().nextInt(this);
+  int get getRandomInt => Random().nextInt(this);
 
   String _formatBytes(int bytes, {required bool binaryPrefixes}) {
     if (bytes <= 0) return 'Unknown Size';
@@ -32,4 +35,10 @@ extension IntExt on int {
 
     return '${value.toIntIfTrue} ${suffixes[i]}';
   }
+}
+
+extension NumExtension on num {
+  num get ln => log(this);
+  num get log2 => log(this) / ln2;
+  num get log10 => log(this) / ln10;
 }